Button Component
Created a new button component.
This component takes a label to display, a type
which can be primary or secondary to alter the
styling, and an optional onClick handler.

Also added two buttons to the homepage for sign up
and login that are wrapped in Link components for
now until we decide what we are doing with them.

Added custom colours to tailwind so we can use
our own theme for the site through tailwind.